refactor: split cli into commands modules
